feat: implement Home and About page layouts with modular section components and styles
