+++ src/pages/About.scss
@@ -1,0 +1,110 @@
+/* About Page Specific Styles */
+
+.about-hero-section {
+  background-color: #fcfcfc;
+  border-bottom: 1px solid #eaeaea;
+  padding: 3.5rem 0;
+}
+
+.about-hero-status {
+  margin-bottom: 1rem;
+  display: inline-block;
+  background-color: rgba(119, 41, 83, 0.08);
+  color: var(--vf-color-accent);
+  border-color: rgba(119, 41, 83, 0.15);
+  font-weight: 700;
+  padding: 0.2rem 0.6rem;
+}
+
+.about-hero-title {
+  color: var(--vf-color-accent);
+  font-weight: 700;
+  font-size: 3rem;
+  line-height: 1.15;
+  margin-bottom: 0.75rem;
+}
+
+.about-hero-lead {
+  color: #555;
+}
+
+.about-hero-icon {
+  font-size: 5rem;
+  color: var(--vf-color-brand);
+  opacity: 0.85;
+  filter: drop-shadow(0 8px 16px rgba(233, 84, 32, 0.15));
+}
+
+.about-intro-section {
+  background-color: #ffffff;
+}
+
+.about-section-title {
+  color: var(--vf-color-accent);
+  font-weight: 700;
+}
+
+.about-values-section {
+  background-color: rgba(233, 84, 32, 0.04);
+  border-top: 1px solid rgba(233, 84, 32, 0.1);
+  border-bottom: 1px solid rgba(233, 84, 32, 0.1);
+  padding: 4rem 0;
+}
+
+.about-values-title {
+  color: var(--vf-color-accent);
+  font-weight: 700;
+  margin-bottom: 2.5rem;
+}
+
+.about-value-col {
+  margin-bottom: 2rem;
+}
+
+.about-value-card {
+  height: 100%;
+  background-color: #ffffff;
+  display: flex;
+  flex-direction: column;
+  padding: 2rem 1.5rem;
+  border-top: 4px solid var(--vf-color-brand);
+  border-radius: 6px;
+  box-shadow: 0 4px 12px rgba(0,0,0,0.02);
+}
+
+.about-value-header {
+  display: flex;
+  align-items: center;
+  gap: 1rem;
+  margin-bottom: 1.25rem;
+}
+
+.about-value-icon-container {
+  display: inline-flex;
+  align-items: center;
+  justify-content: center;
+  width: 52px;
+  height: 52px;
+  border-radius: 50%;
+  background-color: rgba(233, 84, 32, 0.08);
+  flex-shrink: 0;
+}
+
+.about-value-icon {
+  font-size: 1.35rem;
+  color: var(--vf-color-brand);
+}
+
+.about-value-title {
+  margin: 0;
+  color: var(--vf-color-accent);
+  font-weight: 700;
+  font-size: 1.2rem;
+}
+
+.about-value-desc {
+  margin: 0;
+  font-size: 0.9rem;
+  color: #555;
+  line-height: 1.6;
+}
